Refactor aggregator into modular includes
